# GenesisChallenge/Models/SignIn.cs
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.ComponentModel.DataAnnotations.Schema;
using System.Linq;
using System.Threading.Tasks;

namespace GenesisChallenge.Models
{
    public class SignIn
    {
        #region Constractor
        public SignIn()
        {
        }
        #endregion


        #region Properties
        public string Email { get; set; }
        public string Password { get; set; }
       
        #endregion
    }
}